requirements.txt
aiohttp==3.11.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.1
attrs==24.2.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
fake-useragent==1.4.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
gotrue==2.10.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==4.9.3
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.2
openai==0.28.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.0


psycopg2-binary

pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-docx==1.0.1
python-dotenv==1.0.1
pytz==2023.3.post1
realtime==2.0.6
requests==2.31.0
scikit-learn==1.3.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.1
yarl==1.17.2
flask-cors==5.0.0 
psutil==5.9.8
torch==2.2.0
backoff
